minor improvement

+++ includes/class-wp-statistics-install.php
@@ -316,7 +316,7 @@ class Install {
 
 		/**
 		 * Change Charset All Table To New WordPress Collate
-         * Reset Overview Order Meta Box View
+		 * Reset Overview Order Meta Box View
 		 *
 		 * @see https://developer.wordpress.org/reference/classes/wpdb/has_cap/
 		 * @version 13.0.0
@@ -328,7 +328,9 @@ class Install {
 				$wpdb->query( "ALTER TABLE `{$name}` DEFAULT CHARSET=utf8mb4 COLLATE utf8mb4_general_ci ROW_FORMAT = COMPACT;" );
 			}
 		}
-		$wpdb->query("DELETE FROM `{$wpdb->usermeta}` WHERE `meta_key` = 'meta-box-order_toplevel_page_wps_overview_page'");
+		if ( isset( $installed_version ) and version_compare( $installed_version, '13.0', '<=' ) ) {
+			$wpdb->query( "DELETE FROM `{$wpdb->usermeta}` WHERE `meta_key` = 'meta-box-order_toplevel_page_wps_overview_page'" );
+		}
 
 		/**
 		 * Added new Fields to user_online Table
